+++ src/index.ts
@@ -1,3 +1,4 @@
 export * from './config.module';
+export * from './errors';
 export * from './interfaces';
 export * from './services';