+++ src/interfaces/index.ts
@@ -1,1 +1,2 @@
 export * from './config-module-options.interface';
+export * from './configuration-payload.type';